Distinct bytes/string types (refactor, breaks precompiles)
