+++ precompiles/utils/src/data/mod.rs
@@ -14,10 +14,15 @@
 // You should have received a copy of the GNU General Public License
 // along with Moonbeam.  If not, see <http://www.gnu.org/licenses/>.
 
+pub mod bytes;
+pub mod native;
 pub mod xcm;
-pub use affix::paste;
 
+pub use affix::paste;
 pub use alloc::string::String;
+pub use bytes::*;
+pub use native::*;
+
 use {
 	crate::revert::{InjectBacktrace, MayRevert, RevertReason},
 	alloc::borrow::ToOwned,
@@ -28,83 +33,12 @@ use {
 	sp_std::{convert::TryInto, vec, vec::Vec},
 };
 
-type ConstU32Max = ConstU32<{ u32::MAX }>;
-
-/// The `address` type of Solidity.
-/// H160 could represent 2 types of data (bytes20 and address) that are not encoded the same way.
-/// To avoid issues writing H160 is thus not supported.
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub struct Address(pub H160);
-
-impl From<H160> for Address {
-	fn from(a: H160) -> Address {
-		Address(a)
-	}
-}
-
-impl From<Address> for H160 {
-	fn from(a: Address) -> H160 {
-		a.0
-	}
-}
-
-/// The `bytes`/`string` type of Solidity.
-/// It is different from `Vec<u8>` which will be serialized with padding for each `u8` element
-/// of the array, while `Bytes` is tightly packed.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct Bytes(pub Vec<u8>);
-
-impl Bytes {
-	/// Interpret as `bytes`.
-	pub fn as_bytes(&self) -> &[u8] {
-		&self.0
-	}
-
-	/// Interpret as `string`.
-	/// Can fail if the string is not valid UTF8.
-	pub fn as_str(&self) -> Result<&str, sp_std::str::Utf8Error> {
-		sp_std::str::from_utf8(&self.0)
-	}
-}
-
-impl From<&[u8]> for Bytes {
-	fn from(a: &[u8]) -> Self {
-		Self(a.to_owned())
-	}
-}
-
-impl From<&str> for Bytes {
-	fn from(a: &str) -> Self {
-		a.as_bytes().into()
-	}
-}
-
-impl Into<Vec<u8>> for Bytes {
-	fn into(self) -> Vec<u8> {
-		self.0
-	}
-}
-
-/// Same as `Bytes` but with an additional length bound on read.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct BoundedBytes<S> {
-	inner: Vec<u8>,
-	_phantom: PhantomData<S>,
-}
-
-impl<S> BoundedBytes<S> {
-	pub fn into_vec(self) -> Vec<u8> {
-		self.inner
-	}
-}
-
-impl<S> From<Bytes> for BoundedBytes<S> {
-	fn from(from: Bytes) -> Self {
-		Self {
-			inner: from.0,
-			_phantom: PhantomData,
-		}
-	}
+/// Data that can be converted from and to EVM data types.
+pub trait EvmData: Sized {
+	fn read(reader: &mut EvmDataReader) -> MayRevert<Self>;
+	fn write(writer: &mut EvmDataWriter, value: Self);
+	fn has_static_size() -> bool;
+	fn solidity_type() -> String;
 }
 
 /// Wrapper around an EVM input slice, helping to parse it.
@@ -341,400 +275,6 @@ impl EvmDataWriter {
 impl Default for EvmDataWriter {
 	fn default() -> Self {
 		Self::new()
-	}
-}
-
-/// Data that can be converted from and to EVM data types.
-pub trait EvmData: Sized {
-	fn read(reader: &mut EvmDataReader) -> MayRevert<Self>;
-	fn write(writer: &mut EvmDataWriter, value: Self);
-	fn has_static_size() -> bool;
-	fn solidity_type() -> String;
-}
-
-impl EvmData for () {
-	fn read(_reader: &mut EvmDataReader) -> MayRevert<Self> {
-		Ok(())
-	}
-
-	fn write(_writer: &mut EvmDataWriter, _value: Self) {}
-
-	fn has_static_size() -> bool {
-		true
-	}
-
-	fn solidity_type() -> String {
-		String::from("()")
-	}
-}
-
-#[impl_for_tuples(1, 18)]
-impl EvmData for Tuple {
-	fn has_static_size() -> bool {
-		for_tuples!(#( Tuple::has_static_size() )&*)
-	}
-
-	fn read(reader: &mut EvmDataReader) -> MayRevert<Self> {
-		if !Self::has_static_size() {
-			let reader = &mut reader.read_pointer()?;
-			let mut index = 0;
-			Ok(for_tuples!( ( #( {
-				let elem = reader.read::<Tuple>().in_tuple(index)?;
-				index +=1;
-				elem
-			} ),* ) ))
-		} else {
-			let mut index = 0;
-			Ok(for_tuples!( ( #( {
-				let elem = reader.read::<Tuple>().in_tuple(index)?;
-				index +=1;
-				elem
-			} ),* ) ))
-		}
-	}
-
-	fn write(writer: &mut EvmDataWriter, value: Self) {
-		if !Self::has_static_size() {
-			let mut inner_writer = EvmDataWriter::new();
-			for_tuples!( #( Tuple::write(&mut inner_writer, value.Tuple); )* );
-			writer.write_pointer(inner_writer.build());
-		} else {
-			for_tuples!( #( Tuple::write(writer, value.Tuple); )* );
-		}
-	}
-
-	fn solidity_type() -> String {
-		let mut subtypes = Vec::new();
-		for_tuples!( #( subtypes.push(Tuple::solidity_type()); )* );
-		alloc::format!("({})", subtypes.join(","))
-	}
-}
-
-impl EvmData for H256 {
-	fn read(reader: &mut EvmDataReader) -> MayRevert<Self> {
-		let range = reader.move_cursor(32)?;
-
-		let data = reader
-			.input
-			.get(range)
-			.ok_or_else(|| RevertReason::read_out_of_bounds("bytes32"))?;
-
-		Ok(H256::from_slice(data))
-	}
-
-	fn write(writer: &mut EvmDataWriter, value: Self) {
-		writer.data.extend_from_slice(value.as_bytes());
-	}
-
-	fn has_static_size() -> bool {
-		true
-	}
-
-	fn solidity_type() -> String {
-		String::from("bytes32")
-	}
-}
-
-impl EvmData for Address {
-	fn read(reader: &mut EvmDataReader) -> MayRevert<Self> {
-		let range = reader.move_cursor(32)?;
-
-		let data = reader
-			.input
-			.get(range)
-			.ok_or_else(|| RevertReason::read_out_of_bounds("address"))?;
-
-		Ok(H160::from_slice(&data[12..32]).into())
-	}
-
-	fn write(writer: &mut EvmDataWriter, value: Self) {
-		H256::write(writer, value.0.into());
-	}
-
-	fn has_static_size() -> bool {
-		true
-	}
-
-	fn solidity_type() -> String {
-		String::from("address")
-	}
-}
-
-impl EvmData for U256 {
-	fn read(reader: &mut EvmDataReader) -> MayRevert<Self> {
-		let range = reader.move_cursor(32)?;
-
-		let data = reader
-			.input
-			.get(range)
-			.ok_or_else(|| RevertReason::read_out_of_bounds("uint256"))?;
-
-		Ok(U256::from_big_endian(data))
-	}
-
-	fn write(writer: &mut EvmDataWriter, value: Self) {
-		let mut buffer = [0u8; 32];
-		value.to_big_endian(&mut buffer);
-		writer.data.extend_from_slice(&buffer);
-	}
-
-	fn has_static_size() -> bool {
-		true
-	}
-
-	fn solidity_type() -> String {
-		String::from("uint256")
-	}
-}
-
-macro_rules! impl_evmdata_for_uints {
-	($($uint:ty, )*) => {
-		$(
-			impl EvmData for $uint {
-				fn read(reader: &mut EvmDataReader) -> MayRevert<Self> {
-					let value256: U256 = reader.read()
-					.map_err(|_| RevertReason::read_out_of_bounds(
-						Self::solidity_type()
-					))?;
-
-					value256
-						.try_into()
-						.map_err(|_| RevertReason::value_is_too_large(
-							Self::solidity_type()
-						).into())
-				}
-
-				fn write(writer: &mut EvmDataWriter, value: Self) {
-					U256::write(writer, value.into());
-				}
-
-				fn has_static_size() -> bool {
-					true
-				}
-
-				fn solidity_type() -> String {
-					alloc::format!("uint{}", core::mem::size_of::<Self>() * 8)
-				}
-			}
-		)*
-	};
-}
-
-impl_evmdata_for_uints!(u8, u16, u32, u64, u128,);
-
-impl EvmData for bool {
-	fn read(reader: &mut EvmDataReader) -> MayRevert<Self> {
-		let h256 = H256::read(reader).map_err(|_| RevertReason::read_out_of_bounds("bool"))?;
-
-		Ok(!h256.is_zero())
-	}
-
-	fn write(writer: &mut EvmDataWriter, value: Self) {
-		let mut buffer = [0u8; 32];
-		if value {
-			buffer[31] = 1;
-		}
-
-		writer.data.extend_from_slice(&buffer);
-	}
-
-	fn has_static_size() -> bool {
-		true
-	}
-
-	fn solidity_type() -> String {
-		String::from("bool")
-	}
-}
-
-impl EvmData for Bytes {
-	fn read(reader: &mut EvmDataReader) -> MayRevert<Self> {
-		Ok(Bytes(BoundedBytes::<ConstU32Max>::read(reader)?.into_vec()))
-	}
-
-	fn write(writer: &mut EvmDataWriter, value: Self) {
-		BoundedBytes::<ConstU32Max>::write(
-			writer,
-			BoundedBytes {
-				inner: value.0,
-				_phantom: PhantomData,
-			},
-		);
-	}
-
-	fn has_static_size() -> bool {
-		false
-	}
-
-	fn solidity_type() -> String {
-		String::from("bytes")
-	}
-}
-
-impl<S: Get<u32>> EvmData for BoundedBytes<S> {
-	fn read(reader: &mut EvmDataReader) -> MayRevert<Self> {
-		let mut inner_reader = reader.read_pointer()?;
-
-		// Read bytes/string size.
-		let array_size: usize = inner_reader
-			.read::<U256>()
-			.map_err(|_| RevertReason::read_out_of_bounds("length"))?
-			.try_into()
-			.map_err(|_| RevertReason::value_is_too_large("length"))?;
-
-		if array_size > S::get() as usize {
-			return Err(RevertReason::value_is_too_large("length").into());
-		}
-
-		// Get valid range over the bytes data.
-		let range = inner_reader.move_cursor(array_size)?;
-
-		let data = inner_reader
-			.input
-			.get(range)
-			.ok_or_else(|| RevertReason::read_out_of_bounds("bytes/string"))?;
-
-		let bytes = Self {
-			inner: data.to_owned(),
-			_phantom: PhantomData,
-		};
-
-		Ok(bytes)
-	}
-
-	fn write(writer: &mut EvmDataWriter, value: Self) {
-		let value = value.into_vec();
-		let length = value.len();
-
-		// Pad the data.
-		// Leave it as is if a multiple of 32, otherwise pad to next
-		// multiple or 32.
-		let chunks = length / 32;
-		let padded_size = match length % 32 {
-			0 => chunks * 32,
-			_ => (chunks + 1) * 32,
-		};
-
-		let mut value = value.to_vec();
-		value.resize(padded_size, 0);
-
-		writer.write_pointer(
-			EvmDataWriter::new()
-				.write(U256::from(length))
-				.write_raw_bytes(&value)
-				.build(),
-		);
-	}
-
-	fn has_static_size() -> bool {
-		false
-	}
-
-	fn solidity_type() -> String {
-		String::from("bytes")
-	}
-}
-
-impl<T: EvmData> EvmData for Vec<T> {
-	fn read(reader: &mut EvmDataReader) -> MayRevert<Self> {
-		BoundedVec::<T, ConstU32Max>::read(reader).map(|x| x.into_vec())
-	}
-
-	fn write(writer: &mut EvmDataWriter, value: Self) {
-		BoundedVec::<T, ConstU32Max>::write(
-			writer,
-			BoundedVec {
-				inner: value,
-				_phantom: PhantomData,
-			},
-		)
-	}
-
-	fn has_static_size() -> bool {
-		false
-	}
-
-	fn solidity_type() -> String {
-		alloc::format!("{}[]", T::solidity_type())
-	}
-}
-
-/// Wrapper around a Vec that provides a max length bound on read.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct BoundedVec<T, S> {
-	inner: Vec<T>,
-	_phantom: PhantomData<S>,
-}
-
-impl<T, S: Get<u32>> BoundedVec<T, S> {
-	pub fn into_vec(self) -> Vec<T> {
-		self.inner
-	}
-}
-
-impl<T: EvmData, S: Get<u32>> EvmData for BoundedVec<T, S> {
-	fn read(reader: &mut EvmDataReader) -> MayRevert<Self> {
-		let mut inner_reader = reader.read_pointer()?;
-
-		let array_size: usize = inner_reader
-			.read::<U256>()
-			.map_err(|_| RevertReason::read_out_of_bounds("length"))?
-			.try_into()
-			.map_err(|_| RevertReason::value_is_too_large("length"))?;
-
-		if array_size > S::get() as usize {
-			return Err(RevertReason::value_is_too_large("length").into());
-		}
-
-		let mut array = vec![];
-
-		let mut item_reader = EvmDataReader {
-			input: inner_reader
-				.input
-				.get(32..)
-				.ok_or_else(|| RevertReason::read_out_of_bounds("array content"))?,
-			cursor: 0,
-		};
-
-		for i in 0..array_size {
-			array.push(item_reader.read().in_array(i)?);
-		}
-
-		Ok(BoundedVec {
-			inner: array,
-			_phantom: PhantomData,
-		})
-	}
-
-	fn write(writer: &mut EvmDataWriter, value: Self) {
-		let value = value.into_vec();
-		let mut inner_writer = EvmDataWriter::new().write(U256::from(value.len()));
-
-		for inner in value {
-			// Any offset in items are relative to the start of the item instead of the
-			// start of the array. However if there is offseted data it must but appended after
-			// all items (offsets) are written. We thus need to rely on `compute_offsets` to do
-			// that, and must store a "shift" to correct the offsets.
-			let shift = inner_writer.data.len();
-			let item_writer = EvmDataWriter::new().write(inner);
-
-			inner_writer = inner_writer.write_raw_bytes(&item_writer.data);
-			for mut offset_datum in item_writer.offset_data {
-				offset_datum.offset_shift += 32;
-				offset_datum.offset_position += shift;
-				inner_writer.offset_data.push(offset_datum);
-			}
-		}
-
-		writer.write_pointer(inner_writer.build());
-	}
-
-	fn has_static_size() -> bool {
-		false
-	}
-
-	fn solidity_type() -> String {
-		alloc::format!("{}[]", T::solidity_type())
 	}
 }
